% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CONDITION_OCCURRENCE}
\alias{CONDITION_OCCURRENCE}
\title{CONDITION_OCCURRENCE}
\format{
A data frame with 1205 rows and 5 variables:
\describe{
\item{\code{person_id}}{double COLUMN_DESCRIPTION}
\item{\code{gender}}{character COLUMN_DESCRIPTION}
\item{\code{condition_start_date}}{double COLUMN_DESCRIPTION}
\item{\code{condition_end_date}}{double COLUMN_DESCRIPTION}
\item{\code{condition_source}}{character COLUMN_DESCRIPTION}
}
}
\usage{
CONDITION_OCCURRENCE
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
