% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DRUG_EXPOSURE}
\alias{DRUG_EXPOSURE}
\title{DRUG_EXPOSURE}
\format{
A data frame with 10072 rows and 5 variables:
\describe{
\item{\code{person_id}}{double COLUMN_DESCRIPTION}
\item{\code{gender}}{character COLUMN_DESCRIPTION}
\item{\code{drug_exposure_start_date}}{double COLUMN_DESCRIPTION}
\item{\code{drug_exposure_end_date}}{double COLUMN_DESCRIPTION}
\item{\code{drug_source}}{character COLUMN_DESCRIPTION}
}
}
\usage{
DRUG_EXPOSURE
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
