% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MEASUREMENT}
\alias{MEASUREMENT}
\title{MEASUREMENT}
\format{
A data frame with 614 rows and 4 variables:
\describe{
\item{\code{person_id}}{double COLUMN_DESCRIPTION}
\item{\code{gender}}{character COLUMN_DESCRIPTION}
\item{\code{measurement_date}}{double COLUMN_DESCRIPTION}
\item{\code{measurement_source}}{character COLUMN_DESCRIPTION}
}
}
\usage{
MEASUREMENT
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
