% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OBSERVATION}
\alias{OBSERVATION}
\title{OBSERVATION}
\format{
A data frame with 1378 rows and 4 variables:
\describe{
\item{\code{person_id}}{double COLUMN_DESCRIPTION}
\item{\code{gender}}{character COLUMN_DESCRIPTION}
\item{\code{observation_date}}{double COLUMN_DESCRIPTION}
\item{\code{observation_source}}{character COLUMN_DESCRIPTION}
}
}
\usage{
OBSERVATION
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
