% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PROCEDURE_OCCURRENCE}
\alias{PROCEDURE_OCCURRENCE}
\title{PROCEDURE_OCCURRENCE}
\format{
A data frame with 3754 rows and 4 variables:
\describe{
\item{\code{person_id}}{double COLUMN_DESCRIPTION}
\item{\code{gender}}{character COLUMN_DESCRIPTION}
\item{\code{procedure_date}}{double COLUMN_DESCRIPTION}
\item{\code{procedure_source}}{character COLUMN_DESCRIPTION}
}
}
\usage{
PROCEDURE_OCCURRENCE
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
